// packages/web/src/shared/types/api.ts
export interface Embedding {
  id: number
  uri: string
  text: string
  model_name: string
  task_type?: string
  embedding: number[]
  original_content?: string
  converted_format?: string
  created_at: string
  updated_at: string
}